feat(components): adiciona componentes header e footer reutilizaveis
- cria componente Header com navegacao desktop, sidebar tablet e menu mobile FAB
- cria componente Footer com links de navegacao, categorias e redes sociais
- refatora todas as paginas para usar os novos componentes
- remove codigo duplicado de header/footer de cada pagina
- atualiza testes para refletir as mudancas nos componentes
- melhora manutencao e consistencia visual entre paginas
