+++ src/pages/EventDetails.css
@@ -67,13 +67,13 @@
 /* Loading and Error States */
 .loading-container,
 .error-container {
+  flex: 1;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  min-height: 100vh;
   text-align: center;
-  padding: 2rem;
+  padding: 8rem 2rem 2rem;
 }
 
 .loading-container .spinner {
