+++ src/pages/NotFound.css
@@ -1,15 +1,20 @@
 .notfound-page {
   min-height: 100vh;
   display: flex;
-  align-items: center;
-  justify-content: center;
+  flex-direction: column;
   background: linear-gradient(135deg, var(--background) 0%, var(--surface) 100%);
-  padding: 2rem;
 }
 
 .notfound-content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
   text-align: center;
   max-width: 500px;
+  margin: 0 auto;
+  padding: 8rem 2rem 2rem;
 }
 
 .notfound-code {
