+++ src/pages/NotFound.jsx
@@ -1,5 +1,7 @@
 import { useNavigate } from 'react-router-dom'
 import { Home, ArrowLeft } from 'lucide-react'
+import Header from '../components/Header'
+import Footer from '../components/Footer'
 import './NotFound.css'
 
 function NotFound() {
@@ -7,10 +9,12 @@ function NotFound() {
 
   return (
     <div className="notfound-page">
+      <Header />
+
       <div className="notfound-content">
         <div className="notfound-code">404</div>
-        <h1>Página não encontrada</h1>
-        <p>A página que você está procurando não existe ou foi movida.</p>
+        <h1>Pagina nao encontrada</h1>
+        <p>A pagina que voce esta procurando nao existe ou foi movida.</p>
         <div className="notfound-actions">
           <button onClick={() => navigate(-1)} className="notfound-btn secondary">
             <ArrowLeft size={18} />
@@ -18,10 +22,12 @@ function NotFound() {
           </button>
           <button onClick={() => navigate('/')} className="notfound-btn primary">
             <Home size={18} />
-            Ir para Início
+            Ir para Inicio
           </button>
         </div>
       </div>
+
+      <Footer />
     </div>
   )
 }
